# Documents/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "COSC2640 A3 API",
    "description": "The documentation and testing tool for COSC2640 A3 Microservices API.",
    "version": "v1.0"
  },
  "paths": {
...
    "/payment/paypal/{enrolmentId}": {
      "post": {
        "tags": [
          "Payment"
        ],
        "summary": "For student. To make a payment using <b>Paypal</b> for an enrolment's invoice.",
        "description": "Request signature:\r\n\r\n<pre>\rPOST /payment/paypal/{string}\rHeaders\r    \"AccountId\": string\r    \"Authorization\": \"Bearer token\"\rBody\r    {\r        paypalEmail: string,\r        orderId: string,\r        amount: number,\r        authorizationId: string\r    }\r</pre>\r\n\r\n\r\nReturned object signature:\r\n\r\n<pre>\r{\r    id: string,\r    isPaid: boolean,\r    paymentMethod: string,\r    paymentId: string,\r    transactionId: string,\r    paymentStatus: string,\r    paidOn: string\r}\r</pre>\r\n",
        "parameters": [
          {
            "name": "enrolmentId",
            "in": "path",
            "description": "The account's ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The payment authorization data required for capturing money.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaypalAuthorization"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PaypalAuthorization"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PaypalAuthorization"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request was successfully processed."
          },
          "401": {
            "description": "Authorization failed: expired or mismatched or insufficient."
          }
        }
      }
    },
    "/payment/card/{enrolmentId}": {
      "post": {
        "tags": [
          "Payment"
        ],
        "summary": "For student. To make a payment using <b>Bank Card</b> for an enrolment's invoice.",
        "description": "Request signature:\r\n\r\n<pre>\rPOST /payment/card/{string}\rHeaders\r    \"AccountId\": string\r    \"Authorization\": \"Bearer token\"\rBody\r    {\r        cardType: string,\r        last4Digits: string,\r        tokenId: number,\r        details: {\r            classroomId: string,\r            className: string,\r            amount: number\r        }\r    }\r</pre>\r\n\r\n\r\nReturned object signature:\r\n\r\n<pre>\r{\r    id: string,\r    isPaid: boolean,\r    paymentMethod: string,\r    paymentId: string,\r    transactionId: string,\r    chargeId: string\r    paymentStatus: string,\r    paidOn: string\r}\r</pre>\r\n",
        "parameters": [
          {
            "name": "enrolmentId",
            "in": "path",
            "description": "The account's ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The payment authorization data required for capturing money.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StripeAuthorization"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StripeAuthorization"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StripeAuthorization"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request was successfully processed."
          },
          "401": {
            "description": "Authorization failed: expired or mismatched or insufficient."
          }
        }
      }
    },
    "/payment/google/{enrolmentId}": {
      "post": {
        "tags": [
          "Payment"
        ],
        "summary": "For student. To make a payment using <b>Google Pay</b> for an enrolment's invoice.",
        "description": "Request signature:\r\n\r\n<pre>\rPOST /payment/card/{string}\rHeaders\r    \"AccountId\": string\r    \"Authorization\": \"Bearer token\"\rBody\r    {\r        cardType: string,\r        last4Digits: string,\r        tokenId: number,\r        details: {\r            classroomId: string,\r            className: string,\r            amount: number\r        }\r    }\r</pre>\r\n\r\n\r\nReturned object signature:\r\n\r\n<pre>\r{\r    id: string,\r    isPaid: boolean,\r    paymentMethod: string,\r    paymentId: string,\r    transactionId: string,\r    chargeId: string\r    paymentStatus: string,\r    paidOn: string\r}\r</pre>\r\n",
        "parameters": [
          {
            "name": "enrolmentId",
            "in": "path",
            "description": "The account's ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The payment authorization data required for capturing money.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StripeAuthorization"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StripeAuthorization"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StripeAuthorization"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request was successfully processed."
          },
          "401": {
            "description": "Authorization failed: expired or mismatched or insufficient."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PaymentDetail": {
        "type": "object",
        "properties": {
          "classroomId": {
            "type": "string",
            "nullable": true
          },
          "className": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "PaypalAuthorization": {
        "type": "object",
        "properties": {
          "paypalEmail": {
            "type": "string",
            "nullable": true
          },
          "orderId": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "authorizationId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StripeAuthorization": {
        "type": "object",
        "properties": {
          "cardType": {
            "type": "string",
            "nullable": true
          },
          "last4Digits": {
            "type": "string",
            "nullable": true
          },
          "tokenId": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/PaymentDetail"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}